requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.1
certifi==2024.8.30
click==8.1.7
fastapi==0.115.2
filelock==3.13.1
fsspec==2024.2.0
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.3
kornia==0.7.3
kornia_rs==0.1.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
starlette==0.39.2
sympy==1.12
timm==1.0.9
tokenizers==0.20.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.9.0
urllib3==2.2.3
uvicorn==0.31.1
